EA BNK/SCHl: split up distinct formats
